requirements.txt
streamlit
pymongo[srv]
kafka-python
pandas
altair
python-dotenv
